Fixing Linear's code
Linear forgot an increment in the functions which read and parse the
temperature and cell voltage values. From vendor, when using multiple
stacked 6803's, this would cause the controller to read the PEC byte as
a cell voltage or temperature value for the above 6803. This fixes that
and increases that counter. This increase realigns the single
dimensional array read in from SPI with the two dimensional array used
in the general code.